+++ benchmarks/verus/verina/files/verina_advanced_48_task.rs
@@ -8,10 +8,10 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn merge_sort(list: Vec<i32>) -> (result: Vec<i32>)
+fn merge_sort(list: Vec<i8>) -> (result: Vec<i8>)
     ensures
         result.len() == list.len(),
-        forall|i: int, j: int| 0 <= i < j < result.len() ==> result[i] <= result[j],
+        forall|i: int, j: int| 0 <= i < j < result.len() ==> result[i] as int <= result[j] as int,
         result@ =~= list@,
 // </vc-spec>
 // <vc-code>
