+++ benchmarks/verus/verina/files/verina_basic_82_task.rs
@@ -8,12 +8,12 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn remove_front(a: &Vec<i32>) -> (result: Vec<i32>)
+fn remove_front(a: &Vec<i8>) -> (result: Vec<i8>)
     requires a.len() > 0,
     ensures
         a.len() > 0,
         result.len() == a.len() - 1,
-        forall|i: int| 0 <= i < result.len() ==> result[i] == a[i + 1],
+        forall|i: int| 0 <= i < result.len() ==> result[i] as int == a[i + 1] as int,
 // </vc-spec>
 // <vc-code>
 {
