+++ benchmarks/verus/verina/files/verina_basic_96_task.rs
@@ -8,10 +8,10 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn swap_simultaneous(x: i32, y: i32) -> (result: (i32, i32))
+fn swap_simultaneous(x: i8, y: i8) -> (result: (i8, i8))
     ensures
-        result.0 == y,
-        result.1 == x,
+        result.0 as int == y as int,
+        result.1 as int == x as int,
         x != y ==> result.0 != x && result.1 != y,
 // </vc-spec>
 // <vc-code>
